- Reconcile previous commit with master - move cypress utilities out of integration directory - fixed typo in package.json
