+++ cypress/utils/testCategoryHelperUtils.ts
