+++ cypress/utils/testHelperEnums.ts
