+++ cypress/utils/testNotesHelperUtils.ts
